A major update.
Summary:
* Translation source moved toan Excel sheet, the .resx files are generated
* Breaking out common Strings to a section that is identical for all
plugins (the plan is to make this a library, but it may be moved back
to the common file now when the .resx are generated anyway)
* Set up a UnitUtil class that handles labels and in/output of all
units (except TRIMP, that should be handled similarly within that
plugin). Below the interface the ST routines are used. (Unfortunately,
a layer on top of ST is needed). This is not a separate library as I
have not found a simple way to get the reference to the application
from the library.
* All tables and graphs uses the library to get labels and format
ouputs. The presentation is much more uniform.
* As inputs are handled with ST routines, it is possible to specify
the unit when entering, regardless of the system unit
* All internal handling is done using SI units, scaled to presentation
units when displaying only. Simpler and cleaner handling
* More tables are localised, using the ST setting
* Some display update handling. For example Unique Routes plugin
settings required a restart to update presentation.